Migrated VarsHookSet to AppStateProvider
